Fix web app crash: keep expo-media-library out of the web bundle
expo-media-library has no web support in SDK 56 and throws "Cannot find
native module 'ExpoMediaLibraryNext'" at import time, crashing the
deployed web app on load. Split downloadPhoto into a web variant (blob +
anchor) and a native-only module so Metro never bundles the native
import for web.
